Initial Arch Linux support.
Signed-off-by: Julian Phillips <phillips.julian@gmail.com>

[pacman] Adding pacman plugin.

Signed-off-by: Julian Phillips <phillips.julian@gmail.com>

[ArchLinux] Create new ArchPlugin tagging class.

The Arch policy sets this new tagging class as its only valid subclass
and the Pacman plugin now subclasses this new tagging class. With this
commit the Pacman plugin is the only plugin that gets executed on an
Arch system. Support for additional plugins will be added in subsequent
commits, after testing.

Additonal minor changes include the removal of an unused import from
the Arch policy and formatting changes in the Pacman plugin.

Signed-off-by: Julian Phillips <phillips.julian@gmail.com>

Tagged 37 classes with the ArchPlugin class.

All newly tagged plugins are ones that execute as is on an Arch Linux
system. No changes were made to plugin execution code. The ArchPlugin
tagging class is simply imported and appended to the existing list of
derived classes in each plugin.

Signed-off-by: Julian Phillips <phillips.julian@gmail.com>

Tagged 22 more plugins with the ArchPlugin class.

These newly tagged plugins execute as is and produce useful output on
an Arch Linux system without changes to plugin execution code. The
ArchPlugin tagging class is simply imported and appended to the
existing list of derived classes in each plugin. This brings the total
number of classes tagged with the ArchPlugin to 60.

Signed-off-by: Julian Phillips <phillips.julian@gmail.com>